==> Computation/Rustbucket Analyses/ekd40_tmp/03_clustermap.R <==
#!/usr/bin/env R


library(ClusterMap)


# Allen markers file generated by script 02 on rustbucket
# Allen RDS object generated by script 02, renamed/moved to ClusterMap dir

# Avalos marker file generated by ClusterMap script on machine edsouza
# Same script was used to save avalos RDS object, then filezilla moved both to this location in dir ClusterMap


marker_file_list <- c(avalos = 'ClusterMap/avalos.markers.csv', 
		      allen = 'ClusterMap/allen.markers.csv')

avalos <- readRDS('ClusterMap/avalos.RDS')
allen <- readRDS('ClusterMap/allen.RDS') # cp s_o.bythebook.RDS ClusterMap/allen.RDS

single_obj_list <- c(avalos = avalos, allen = allen)

res <- cluster_map(marker_file_list, edge_cutoff = 0.1, 
		                      output = 'ClusterMap/AvalosAllen', 
				                         single_obj_list = single_obj_list)
res <- cluster_map(marker_file_list, edge_cutoff = 0.1, output = 'ClusterMap/AvalosAllen')
res
write.csv(res, 'ClusterMap/res-AvalosAllen.csv')


###########################################################################

# Now the same, with Zinzen + Avalos
rm(list=ls()); gc()

marker_file_list <- c(avalos = 'ClusterMap/avalos.markers.csv', 
		      zinzen = 'ClusterMap/zinzen.markers.csv')

avalos <- readRDS('ClusterMap/avalos.RDS')
zinzen <- readRDS('ClusterMap/zinzen.RDS'

single_obj_list <- c(avalos = avalos, zinzen = zinzen)

res <- cluster_map(marker_file_list, edge_cutoff = 0.1, 
		                      output = 'ClusterMap/ZinzenAvalos', 
				                         single_obj_list = single_obj_list)
res <- cluster_map(marker_file_list, edge_cutoff = 0.1, output = 'ClusterMap/ZinzenAvalos')
res
write.csv(res, 'ClusterMap/res-ZinzenAvalos.csv')
